# New vectorised and dependency-less `noise` implementation