--- a/examples/random_lines/sketch_random_lines.py
+++ b/examples/random_lines/sketch_random_lines.py
@@ -15,16 +15,12 @@ class RandomLinesSketch(vsketch.Vsketch):
 
         x_coords = np.linspace(0, 25, 1000)
 
+        perlin = self.noise(
+            x_coords * self.x_freq, np.arange(self.num_line) / self.num_line * self.y_freq
+        )
+
         for i in range(self.num_line):
-            y_coords = (
-                np.array(
-                    [
-                        self.noise(x * self.x_freq, i / self.num_line * self.y_freq)
-                        for x in x_coords
-                    ]
-                )
-                + self.y_offset / self.num_line * i
-            )
+            y_coords = perlin[:, i] + self.y_offset / self.num_line * i
             self.polygon(x_coords, y_coords)
 
     def finalize(self) -> None: