--- a/mypy.ini
+++ b/mypy.ini
@@ -2,5 +2,5 @@
 ignore_missing_imports = True
 allow_redefinition = True
 disallow_incomplete_defs = True
-files = vsketch,vsketch_cli,tests
+files = vsketch,vsketch_cli,examples,tests