--- a/pyproject.toml
+++ b/pyproject.toml
@@ -59,6 +59,7 @@ recommonmark = { version = ">=0.6,<0.8", optional = true }
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.1"
+pytest-benchmark = "^3.2.3"
 mypy = "^0.800"
 jupyter = "^1.0.0"
 nbconvert = "5.6.1"