--- a/tests/test_random.py
+++ b/tests/test_random.py
@@ -1,3 +1,5 @@
+import numbers
+
 import vsketch
 
 
@@ -63,3 +65,16 @@ def test_noise_seed(vsk):
     assert n1 != n4
     assert n2 != n4
     assert n1 == n3
+
+
+def test_noise_dimensions(vsk):
+    assert isinstance(vsk.noise(1.5), numbers.Number)
+
+    assert vsk.noise([0, 1]).shape == (2,)
+    assert vsk.noise([0, 1], 3).shape == (2,)
+    assert vsk.noise([0, 1], [2, 3, 4]).shape == (2, 3)
+    assert vsk.noise([0, 1], [2, 3, 4], 4).shape == (2, 3)
+    assert vsk.noise([0, 1], [2, 3, 4], [5, 6, 7, 8]).shape == (2, 3, 4)
+    assert vsk.noise([0, 1], 4, [5, 6, 7, 8]).shape == (2, 4)
+    assert vsk.noise(4, 4, [5, 6, 7, 8]).shape == (4,)
+    assert vsk.noise(4, [0, 1], [5, 6, 7, 8]).shape == (2, 4)