--- a/tests/test_random.py
+++ b/tests/test_random.py
@@ -1,6 +1,9 @@
 import numbers
 
+import numpy as np
+
 import vsketch
+from vsketch.noise import Noise
 
 
 def test_random(vsk):
@@ -78,3 +81,22 @@ def test_noise_dimensions(vsk):
     assert vsk.noise([0, 1], 4, [5, 6, 7, 8]).shape == (2, 4)
     assert vsk.noise(4, 4, [5, 6, 7, 8]).shape == (4,)
     assert vsk.noise(4, [0, 1], [5, 6, 7, 8]).shape == (2, 4)
+
+
+def test_noise_benchmark_single(benchmark):
+    noise = Noise()
+    benchmark(noise.perlin, 1.0, 2.0, 3.0)
+
+
+def test_noise_benchmark_small_array(benchmark):
+    noise = Noise()
+    benchmark(
+        noise.perlin, np.linspace(0, 1, 10), np.linspace(0, 1, 10), np.linspace(0, 1, 10)
+    )
+
+
+def test_noise_benchmark_big_array(benchmark):
+    noise = Noise()
+    benchmark(
+        noise.perlin, np.linspace(0, 1, 100), np.linspace(0, 1, 100), np.linspace(0, 1, 100)
+    )