--- a/vsketch/noise.py
+++ b/vsketch/noise.py
@@ -1,4 +1,5 @@
 import math
+from numbers import Number
 from typing import Sequence, Union
 
 import numpy as np
@@ -20,11 +21,12 @@ class Noise:
         rng = np.random.default_rng(seed)
         self._perlin = rng.random(PERLIN_SIZE + 1)
 
+    # @profile
     def perlin(
         self,
-        x: Union[float, Sequence[float]],
-        y: Union[float, Sequence[float]],
-        z: Union[float, Sequence[float]],
+        x: Union[Number, Sequence[Number]],
+        y: Union[Number, Sequence[Number]],
+        z: Union[Number, Sequence[Number]],
     ) -> np.ndarray:
         """Compute perlin noise for a range of values.
 
@@ -33,9 +35,17 @@ class Noise:
         Numpy array of shape (len(x), len(y), len(z)). If any of the input are scalar, they are
         interpreted as a length-1 array.
         """
-        grid = np.abs(np.array(np.meshgrid(x, y, z, indexing="ij", copy=False), dtype=float))
-        grid_i = np.floor(grid).astype(int)
-        grid_f = grid - grid_i
+
+        if isinstance(x, Number) and isinstance(y, Number) and isinstance(z, Number):
+            grid = np.array([x, y, z])
+            single = True
+        else:
+            grid = np.array(np.meshgrid(x, y, z, indexing="ij", copy=False), dtype=float)
+            single = False
+
+        np.abs(grid, out=grid)
+        grid_i = grid.astype(int)
+        grid = grid - grid_i  # faster than in place for small matrices
 
         r = np.zeros(shape=grid.shape[1:])
         ampl = 0.5
@@ -47,19 +57,19 @@ class Noise:
                 + (grid_i[2, ...] << PERLIN_ZWRAPB)
             )
 
-            r_f = 0.5 * (1.0 - np.cos(grid_f * math.pi))
+            r_f = 0.5 * (1.0 - np.cos(grid * math.pi))
 
-            n1 = self._perlin[of & PERLIN_SIZE]
-            n1 += r_f[0, ...] * (self._perlin[(of + 1) & PERLIN_SIZE] - n1)
-            n2 = self._perlin[(of + PERLIN_YWRAP) & PERLIN_SIZE]
-            n2 += r_f[0, ...] * (self._perlin[(of + PERLIN_YWRAP + 1) & PERLIN_SIZE] - n2)
+            n1 = self._perlin[of % PERLIN_SIZE]
+            n1 += r_f[0, ...] * (self._perlin[(of + 1) % PERLIN_SIZE] - n1)
+            n2 = self._perlin[(of + PERLIN_YWRAP) % PERLIN_SIZE]
+            n2 += r_f[0, ...] * (self._perlin[(of + PERLIN_YWRAP + 1) % PERLIN_SIZE] - n2)
             n1 += r_f[1, ...] * (n2 - n1)
 
             of += PERLIN_ZWRAP
-            n2 = self._perlin[of & PERLIN_SIZE]
-            n2 += r_f[0, ...] * (self._perlin[(of + 1) & PERLIN_SIZE] - n2)
-            n3 = self._perlin[(of + PERLIN_YWRAP) & PERLIN_SIZE]
-            n3 += r_f[0, ...] * (self._perlin[(of + PERLIN_YWRAP + 1) & PERLIN_SIZE] - n3)
+            n2 = self._perlin[of % PERLIN_SIZE]
+            n2 += r_f[0, ...] * (self._perlin[(of + 1) % PERLIN_SIZE] - n2)
+            n3 = self._perlin[(of + PERLIN_YWRAP) % PERLIN_SIZE]
+            n3 += r_f[0, ...] * (self._perlin[(of + PERLIN_YWRAP + 1) % PERLIN_SIZE] - n3)
             n2 += r_f[1, ...] * (n3 - n2)
 
             n1 += r_f[2, ...] * (n2 - n1)
@@ -68,13 +78,13 @@ class Noise:
             ampl *= self.amp_falloff
 
             grid_i <<= 1
-            grid_f *= 2
+            grid *= 2
 
-            idx = grid_f >= 1.0
+            idx = grid >= 1.0
             grid_i[idx] += 1
-            grid_f[idx] -= 1.0
+            grid[idx] -= 1.0
 
-        return r
+        return r.reshape(1, 1, 1) if single else r
 
 
 def main():
@@ -97,8 +107,29 @@ def main():
     plt.show()
 
 
+def main_old():
+    import matplotlib.pyplot as plt
+
+    n = Noise()
+
+    for j in range(30):
+        x = np.arange(100)
+        y = np.empty(100)
+
+        for i in range(100):
+            y[i] = n.perlin(i * 0.1, j * 0.1, 0)
+
+        plt.plot(x * 5, j * 10 + 100 * y, "-b")
+
+    plt.axis("square")
+    plt.xlim(0, 640)
+    plt.ylim(0, 360)
+
+    plt.show()
+
+
 if __name__ == "__main__":
-    main()
+    main_old()
 
     # print(Noise().noise_numpy(1, 0, 0))
     # import matplotlib.pyplot as plt