#### Description

The goal of this PR is two-fold:
- remove the dependency on the `noise` library, which is no longer supported and cause installation issues
- make `vsketch.noise()` vectorised for a very significant performance improvement.

The implementation is now in Python + Numpy. The API is the same but now accepts 1D arrays for any or all of the `x`, `y`, and `z` arguments.

The resulting `vsketch.noise()` function is now *much* slower than the previous implementation (which was in Fortran) for a single value, but *much* faster when generating a vectorised output (see the benchmarks below). 

#### Examples

```python
>>> vsk.noise(12.)  # returns a float
0.5713948646260701

>>> vsk.noise(12., 4., 10.)  # perlin space of up to 3 dimensions can be sampled
0.27149380480954677

>>> vsk.noise([0, 0.1, 0.2, 0.3, 0.4])  # x is now an array
array([0.73779253, 0.7397108 , 0.73590816, 0.72425246, 0.69773313])

>>> vsk.noise(np.linspace(0., 1., 10), np.linspace(0., 1., 5))  # x and y are arrays, result is a 2D matrix
array([[0.73779253, 0.61588815, 0.52075717, 0.48219902, 0.50484146],
       [0.74000394, 0.60852933, 0.52967462, 0.4903986 , 0.48789186],
       [0.7342293 , 0.7007292 , 0.53910934, 0.42320587, 0.4755026 ],
       [0.7119145 , 0.72482773, 0.60064276, 0.47560926, 0.47598847],
       [0.68652505, 0.74806387, 0.70442587, 0.55913415, 0.47530898],
       [0.67534299, 0.74523211, 0.74932192, 0.60766379, 0.50194793],
       [0.65110897, 0.67265303, 0.73858756, 0.67539248, 0.54396967],
       [0.6517441 , 0.63630075, 0.70270696, 0.73336312, 0.6288589 ],
       [0.59085755, 0.67609827, 0.73308901, 0.74057962, 0.75528381],
       [0.59373785, 0.67100004, 0.75857507, 0.773705  , 0.78103623]])

>>> vsk.noise(np.linspace(0., 1., 100), np.linspace(0., 1., 50), [0, 100, 200, 300]).shape
(100, 50, 4)
```

#### Benchmark

Are are the two functions being benchmarked, adapted by the `random_flowers` example:

```python
import math
import numpy as np
import vsketch

POINT_PER_LINE = 500
NUM_LINE = 200
RDIR_RANGE = math.pi / 6


def flower_new(vsk: vsketch.Vsketch):
    vsk.size("a4", landscape=True)
    vsk.scale("cm")
    vsk.rotate(-90, degrees=True)

    noise_coord = np.linspace(0, 1, POINT_PER_LINE)
    dirs = np.linspace(0, 2 * math.pi, NUM_LINE)
    perlin = vsk.noise(noise_coord, dirs, [0, 100])

    for i, direction in enumerate(dirs):
        rdir = vsk.map(perlin[:, i, 0], 0, 1, direction - RDIR_RANGE, direction + RDIR_RANGE)
        roffset = vsk.map(perlin[:, i, 1], 0, 1, 0.05, 0.12)
        xoffset = roffset * np.cos(rdir)
        yoffset = roffset * np.sin(rdir)
        vsk.polygon(np.cumsum(xoffset), np.cumsum(yoffset))


def flower_old(vsk: vsketch.Vsketch):
    vsk.size("a4", landscape=True)
    vsk.scale("cm")
    vsk.rotate(-90, degrees=True)

    noise_coord = np.linspace(0, 1, POINT_PER_LINE)
    dirs = np.linspace(0, 2 * math.pi, NUM_LINE)
    for direction in dirs:
        rdir = vsk.map(
            np.array([vsk.noise(x, direction) for x in noise_coord]), 0, 1, direction - RDIR_RANGE, direction + RDIR_RANGE
        )
        roffset = vsk.map(
            np.array([vsk.noise(x, direction, 100) for x in noise_coord]), 0, 1, 0.05, 0.12
        )

        xoffset = roffset * np.cos(rdir)
        yoffset = roffset * np.sin(rdir)

        vsk.polygon(np.cumsum(xoffset), np.cumsum(yoffset))
```

This is a benchmark that uses 10k noise value and includes other processing (such as the `vsk.polygon()` call).

|  | Results | 
| --- | --- |
| Old implementation | 271 ms ± 5.91 ms |
| New implementation | 11.3 s ± 3.04 s |
| New implementation (vectorised) | 49.6 ms ± 1.34 ms |

This benchmark aims at comparing the raw performance with 10M points:

```python
POINT_PER_LINE = 10000
NUM_LINE = 1000

def raw_perf_new():
    vsk = vsketch.Vsketch()
    vsk.noise(range(POINT_PER_LINE), range(NUM_LINE))


def raw_perf_old():
    vsk = vsketch.Vsketch()
    for j in range(POINT_PER_LINE):
        for i in range(NUM_LINE):
            vsk.noise(i, j)
```

|  | Results | 
| --- | --- |
| Old implementation | 10 s ± 324 ms |
| New implementation (vectorised) | 7.06 s ± 280 ms |

Observations:
- Python loops are *that* slow (if what you do inside them is fast)
- the `noise` library is pretty fast (too bad it's not maintained)


#### Checklist

- [x] feature/fix implemented
- [x] `mypy` returns no error
- [x] tests added/updated and `pytest --runslow` succeeds
- [x] documentation added/updated and building with no error (`make clean && make html` in `docs/`)
- [x] examples added/updated
- [x] code formatting ok (`black` and `isort`)